Add readme + Cleanup and updated to match javascript starter
