+++ typescript/IPlayerResult.ts
@@ -1,0 +1,6 @@
+import { IPlayer } from "bomberjam-backend";
+
+export interface IPlayerResult {
+    fitness: number;
+    score: number;
+}
