--- typescript/last-bot/model.json
@@ -1,1 +1,0 @@
-{"modelTopology":{"class_name":"Sequential","config":{"name":"sequential_70","layers":[{"class_name":"Conv2D","config":{"filters":32,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"kernel_regularizer":null,"kernel_constraint":null,"kernel_size":[3,3],"strides":[1,1],"padding":"valid","data_format":"channels_first","dilation_rate":[1,1],"activation":"relu","use_bias":true,"bias_initializer":{"class_name":"Zeros","config":{}},"bias_regularizer":null,"activity_regularizer":null,"bias_constraint":null,"name":"conv2d_Conv2D205","trainable":true,"batch_input_shape":[null,13,13,11],"dtype":"float32"}},{"class_name":"Dropout","config":{"rate":0.15,"noise_shape":null,"seed":null,"name":"dropout_Dropout205","trainable":true}},{"class_name":"Conv2D","config":{"filters":64,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"kernel_regularizer":null,"kernel_constraint":null,"kernel_size":[3,3],"strides":[1,1],"padding":"valid","data_format":"channels_last","dilation_rate":[1,1],"activation":"relu","use_bias":true,"bias_initializer":{"class_name":"Zeros","config":{}},"bias_regularizer":null,"activity_regularizer":null,"bias_constraint":null,"name":"conv2d_Conv2D206","trainable":true}},{"class_name":"Dropout","config":{"rate":0.15,"noise_shape":null,"seed":null,"name":"dropout_Dropout206","trainable":true}},{"class_name":"Conv2D","config":{"filters":128,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"kernel_regularizer":null,"kernel_constraint":null,"kernel_size":[3,3],"strides":[1,1],"padding":"valid","data_format":"channels_last","dilation_rate":[1,1],"activation":"relu","use_bias":true,"bias_initializer":{"class_name":"Zeros","config":{}},"bias_regularizer":null,"activity_regularizer":null,"bias_constraint":null,"name":"conv2d_Conv2D207","trainable":true}},{"class_name":"Dropout","config":{"rate":0.15,"noise_shape":null,"seed":null,"name":"dropout_Dropout207","trainable":true}},{"class_name":"Flatten","config":{"name":"flatten_Flatten69","trainable":true}},{"class_name":"Dense","config":{"units":64,"activation":"relu","use_bias":true,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"bias_initializer":{"class_name":"Zeros","config":{}},"kernel_regularizer":null,"bias_regularizer":null,"activity_regularizer":null,"kernel_constraint":null,"bias_constraint":null,"name":"dense_Dense205","trainable":true}},{"class_name":"Dense","config":{"units":32,"activation":"relu","use_bias":true,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"bias_initializer":{"class_name":"Zeros","config":{}},"kernel_regularizer":null,"bias_regularizer":null,"activity_regularizer":null,"kernel_constraint":null,"bias_constraint":null,"name":"dense_Dense206","trainable":true}},{"class_name":"Dense","config":{"units":6,"activation":"softmax","use_bias":true,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"bias_initializer":{"class_name":"Zeros","config":{}},"kernel_regularizer":null,"bias_regularizer":null,"activity_regularizer":null,"kernel_constraint":null,"bias_constraint":null,"name":"dense_Dense207","trainable":true}}]},"keras_version":"tfjs-layers 1.3.1","backend":"tensor_flow.js"},"weightsManifest":[{"paths":["weights.bin"],"weights":[{"name":"conv2d_Conv2D205/kernel","shape":[3,3,13,32],"dtype":"float32"},{"name":"conv2d_Conv2D205/bias","shape":[32],"dtype":"float32"},{"name":"conv2d_Conv2D206/kernel","shape":[3,3,9,64],"dtype":"float32"},{"name":"conv2d_Conv2D206/bias","shape":[64],"dtype":"float32"},{"name":"conv2d_Conv2D207/kernel","shape":[3,3,64,128],"dtype":"float32"},{"name":"conv2d_Conv2D207/bias","shape":[128],"dtype":"float32"},{"name":"dense_Dense205/kernel","shape":[25088,64],"dtype":"float32"},{"name":"dense_Dense205/bias","shape":[64],"dtype":"float32"},{"name":"dense_Dense206/kernel","shape":[64,32],"dtype":"float32"},{"name":"dense_Dense206/bias","shape":[32],"dtype":"float32"},{"name":"dense_Dense207/kernel","shape":[32,6],"dtype":"float32"},{"name":"dense_Dense207/bias","shape":[6],"dtype":"float32"}]}],"format":"layers-model","generatedBy":"TensorFlow.js tfjs-layers v1.3.1","convertedBy":null}
