+++ typescript/package.json
@@ -4,13 +4,16 @@
   "description": "",
   "main": "bot.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "evolve": "npx ts-node evolve.ts"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@tensorflow/tfjs-node": "1.3.1",
-    "bomberjam-backend": "^1.0.9",
+    "bomberjam-backend": "^1.0.13"
+  },
+  "devDependencies": {
+    "ts-node": "^8.5.4",
     "typescript": "^3.7.2"
   }
 }
