+++ typescript/saved-models/base-model/model.json
@@ -1,0 +1,1 @@
+{"modelTopology":{"class_name":"Sequential","config":{"name":"sequential_24","layers":[{"class_name":"Conv2D","config":{"filters":32,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"kernel_regularizer":null,"kernel_constraint":null,"kernel_size":[3,3],"strides":[1,1],"padding":"valid","data_format":"channels_first","dilation_rate":[1,1],"activation":"relu","use_bias":true,"bias_initializer":{"class_name":"Zeros","config":{}},"bias_regularizer":null,"activity_regularizer":null,"bias_constraint":null,"name":"conv2d_Conv2D45","trainable":true,"batch_input_shape":[null,4,13,11],"dtype":"float32"}},{"class_name":"Dropout","config":{"rate":0.15,"noise_shape":null,"seed":null,"name":"dropout_Dropout45","trainable":true}},{"class_name":"MaxPooling2D","config":{"pool_size":[2,2],"padding":"valid","strides":[2,2],"data_format":"channels_first","name":"max_pooling2d_MaxPooling2D23","trainable":true}},{"class_name":"Conv2D","config":{"filters":128,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"kernel_regularizer":null,"kernel_constraint":null,"kernel_size":[3,3],"strides":[1,1],"padding":"valid","data_format":"channels_first","dilation_rate":[1,1],"activation":"relu","use_bias":true,"bias_initializer":{"class_name":"Zeros","config":{}},"bias_regularizer":null,"activity_regularizer":null,"bias_constraint":null,"name":"conv2d_Conv2D46","trainable":true}},{"class_name":"Dropout","config":{"rate":0.15,"noise_shape":null,"seed":null,"name":"dropout_Dropout46","trainable":true}},{"class_name":"Flatten","config":{"name":"flatten_Flatten23","trainable":true}},{"class_name":"Dense","config":{"units":128,"activation":"relu","use_bias":true,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"bias_initializer":{"class_name":"Zeros","config":{}},"kernel_regularizer":null,"bias_regularizer":null,"activity_regularizer":null,"kernel_constraint":null,"bias_constraint":null,"name":"dense_Dense67","trainable":true}},{"class_name":"Dense","config":{"units":64,"activation":"relu","use_bias":true,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"bias_initializer":{"class_name":"Zeros","config":{}},"kernel_regularizer":null,"bias_regularizer":null,"activity_regularizer":null,"kernel_constraint":null,"bias_constraint":null,"name":"dense_Dense68","trainable":true}},{"class_name":"Dense","config":{"units":6,"activation":"softmax","use_bias":true,"kernel_initializer":{"class_name":"VarianceScaling","config":{"scale":1,"mode":"fan_avg","distribution":"normal","seed":null}},"bias_initializer":{"class_name":"Zeros","config":{}},"kernel_regularizer":null,"bias_regularizer":null,"activity_regularizer":null,"kernel_constraint":null,"bias_constraint":null,"name":"dense_Dense69","trainable":true}}]},"keras_version":"tfjs-layers 1.3.1","backend":"tensor_flow.js"},"weightsManifest":[{"paths":["weights.bin"],"weights":[{"name":"conv2d_Conv2D45/kernel","shape":[3,3,4,32],"dtype":"float32"},{"name":"conv2d_Conv2D45/bias","shape":[32],"dtype":"float32"},{"name":"conv2d_Conv2D46/kernel","shape":[3,3,32,128],"dtype":"float32"},{"name":"conv2d_Conv2D46/bias","shape":[128],"dtype":"float32"},{"name":"dense_Dense67/kernel","shape":[768,128],"dtype":"float32"},{"name":"dense_Dense67/bias","shape":[128],"dtype":"float32"},{"name":"dense_Dense68/kernel","shape":[128,64],"dtype":"float32"},{"name":"dense_Dense68/bias","shape":[64],"dtype":"float32"},{"name":"dense_Dense69/kernel","shape":[64,6],"dtype":"float32"},{"name":"dense_Dense69/bias","shape":[6],"dtype":"float32"}]}],"format":"layers-model","generatedBy":"TensorFlow.js tfjs-layers v1.3.1","convertedBy":null}
